Spotify command: ApplicationCommand-only
